broke up world_editor into mulitple files
